pyproject.toml
[tool.poetry]
name = "tp-app-fastapi"
version = "0.1.0"
description = ""
authors = ["Ruslan <khrg1993@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
sqladmin = "^0.16.1"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
